Style update (CMD check note)
Mostly changed values for functions from dplyr and ggplot2 to strings 
and non-standard evaluations to resolve notes arising from 
quasiquotation and other unbound variables. Other approaches to arrive 
at the same goal was modifying a function slightly to use non-standard 
evaluation, changing the name of a variable in a foreach call, setting a 
default value for a variable and providing '::' in a function call.