Extensive vignette update (R CMD check warning and BiocCheck error)
The installed package size was too large (> 5mb) and the rendered 
vignette directory in the tarball (notame/inst/doc) exceeded the maximum 
of 1mb. Compressing images was not enough. I settled for including two 
vignettes, introduction.Rmd and project_example.Rmd which were rather 
extensively modified from the originals of the same names. In 
introduction.Rmd, I guide the user to the package according to 
Bioconductor instructions, emphasizing big-picture functionality. In 
project_example.Rmd, I demonstrate the functionality in unison with the 
notame protocol article. I aimed to make project_example.Rmd brief so as 
to give the impression of a project template. 


The size of the package is now 2.5 Mb, so a few more, but not all, 
vignettes could be included in the package. It may be better to include 
and extend the  now leftover vignettes (feature_clustering.Rmd, 
MetaboSet.Rmd, preprocessing.Rmd, project_template.Rmd, statistics.Rmd, 
visualizations.Rmd) on a future pkgdown website for notame. 

There is now ample space for a parallel TSE translation of the package.