Fix documentation and namespace
Made documentation line lengths < 80 characters, consolidated the style 
of the description and value sections. Fixed faulty mentions of default 
values. Added citation for some functionality. Fixed some @seealso tags. 
Made notame-package.Rd to facilitate an overview of the package.

Having learned more about namespace, I settled for using :: instead of 
@importFrom for readability of code, except in cases where the functions 
are needed package-wide, in which case I included them at the beginning 
of the utils file (such as various Biobase functions and and ggplot2). 
Supposedly, loading the function using :: takes a few milliseconds in 
comparison to importing the function to namespace, but this is 
negligible.

Now all suggests are functions which have both requirenamespace and 
citation (which produce numerical results or where computation is 
involved in plot), or are used for development (testthat, for example). 
This also reflects (from what I understand primary motivation of the 
suggests field), that packages in suggests are not installed when notame 
is installed. I thought this is a useful guideline for future 
developers. Calls to requireNamespace were removed for packages moved 
from Suggests to Imports. I also removed some dependencies which are no 
longer needed. 

Various small changes to vignette text and documentation to reflect the 
conceptualization of the package. Much of this commit was not required 
by Bioconductor but I thought it would be clear for future developers 
and maintenance.