Commit message:
Remove and de-export functions

In comparison to the release version of notame (0.3.1), I've now removed
- dobc (has been deprecated for a long time)
- mark_qcs (is not used internally, isn't core functionality)
- install_dependencies (not needed when MUVR and batchCorr are in 
CRAN/Bioconductor + the submission guide states that package code 
shouldn't install anything)
- merge_batches (deprecated for a long time)

De-exported: 
- prop_found (doesn't use MetaboSet, not core functionality)
- prop_na (doesn't use MetaboSet, not core functionality)
- merge_exprs ("can be useful")
- finite_helpers (results are better acquired using summary_stats)
- perform_paired_t_test (perform_pairwise_t_test offers same 
functionality 
- dc_cubic_spline, inspect_dc (the utility of these low-level functions 
is not obvious)
- feature clustering functions except the wrapper correct_drift (they 
take a disproportionate amount of real estate, can of course still be 
used with :::)

notame is a large package, and I think removing and de-exporting as 
above to retain core functionality makes the package easier to grasp and 
maintain. The Bioconductor-release is a good excuse/milestone. 

Also, removed @seealso accordingly and some small documentation changes 
to fix faulty links in the previous commit.